requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

asgiref==3.6.0
channels==4.0.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
